maker and taker volume is u128

+++ packages/deepbook/sources/state/trade_params.move
@@ -28,9 +28,9 @@ module deepbook::trade_params {
     public(package) fun taker_fee_for_user(
         self: &TradeParams,
         active_stake: u64,
-        volume_in_deep: u64,
+        volume_in_deep: u128,
     ): u64 {
-        if (active_stake >= self.stake_required && volume_in_deep >= self.stake_required) {
+        if (active_stake >= self.stake_required && volume_in_deep >= (self.stake_required as u128)) {
             self.taker_fee / 2
         } else {
             self.taker_fee
